proper casts
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@3000 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/parsing/eiffel/encode.c
@@ -16,8 +16,8 @@
 #define ENCODE_LENGTH 7
 
 #define FEAT_NAME_FLAG		0
-#define ROUT_TABLE_FLAG		((1<<30)+ (1<<29))
-#define TYPE_TABLE_FLAG		((1<<31) + (1<<29))
+#define ROUT_TABLE_FLAG		(((uint32)1<<30) + ((uint32)1<<29))
+#define TYPE_TABLE_FLAG		(((uint32)1<<31) + ((uint32)1<<29))
 
 /*
  * Function declarations.
